更新百度SearchWithRAG单元测试BUG
